+++ tests/test_type_inference.py
@@ -1,7 +1,8 @@
 import pytest
 
 from visions import EmailAddress
-from visions.test.series import get_geometry_series, get_series
+from visions.test.series import get_series
+from visions.test.series_geometry import get_geometry_series
 from visions.types import Generic
 from visions.typesets import CompleteSet
