+++ tests/typesets/test_complete_set.py
@@ -10,7 +10,8 @@ import pytest
 from shapely import wkt
 
 from visions import CompleteSet
-from visions.test.series import get_geometry_series, get_series
+from visions.test.series import get_series
+from visions.test.series_geometry import get_geometry_series
 from visions.test.utils import (
     cast,
     contains,
