+++ tests/typesets/test_standard_set.py
@@ -3,7 +3,8 @@ import pandas as pd
 import pytest
 
 from visions import StandardSet
-from visions.test.series import get_geometry_series, get_series
+from visions.test.series import get_series
+from visions.test.series_geometry import get_geometry_series
 from visions.test.utils import (
     cast,
     contains,
