+++ tests/typesets/test_standard_set_sparse.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import pytest
 
-from tests.series_sparse import get_sparse_series
 from visions import StandardSet
+from visions.test.series_sparse import get_sparse_series
 from visions.test.utils import contains, get_contains_cases, get_inference_cases, infers
 from visions.types import (
     Boolean,
